+++ src/main/java/me/axlerogue/mysteriumfateboxes/handlers/DrownedHordeHandler.java
@@ -36,5 +36,7 @@ public class DrownedHordeHandler {
                 level.addFreshEntity(drowned);
             }
         }
+        
+        HordeTitleHandler.sendHordeTitle(player);
     }
 }