+++ src/main/java/me/axlerogue/mysteriumfateboxes/handlers/HuskHordeHandler.java
@@ -36,5 +36,7 @@ public class HuskHordeHandler {
                 level.addFreshEntity(husk);
             }
         }
+        
+        HordeTitleHandler.sendHordeTitle(player);
     }
 }