+++ src/main/java/me/axlerogue/mysteriumfateboxes/handlers/RainingFlamingArrowsHandler.java
@@ -37,5 +37,7 @@ public class RainingFlamingArrowsHandler {
                 level.addFreshEntity(arrow);
             }
         }
+        
+        TrapTitleHandler.sendTrapTitle(player);
     }
 }