+++ src/main/java/me/axlerogue/mysteriumfateboxes/handlers/WitherBossHandler.java
@@ -22,5 +22,7 @@ public class WitherBossHandler {
                 level.addFreshEntity(wither);
             }
         }
+        
+        WitherBossTitleHandler.sendWitherTitle(player);
     }
 }