+++ src/main/java/me/axlerogue/mysteriumfateboxes/handlers/ZombieHordeHandler.java
@@ -36,5 +36,7 @@ public class ZombieHordeHandler {
                 level.addFreshEntity(zombie);
             }
         }
+        
+        HordeTitleHandler.sendHordeTitle(player);
     }
 }